Added Zomato Food Delivery Project
The newly started companies are not able to decide the cost that would happen per two people for once. So we can predict the cost per two customers for one time so that the newly started  restaurants and upcoming restaurants will be well prepared how the restaurant should invest in improving the ambiance and all other stuff to attract the customers. Assume we are the analyst team that Zomato has organized to help new and upcoming restaurants by letting them know the various reasons that customers look for and build a model which able to predict the cost for two people.